+++ refinedstorage-neoforge/src/main/java/com/refinedmods/refinedstorage/neoforge/ConfigImpl.java
@@ -411,7 +411,7 @@ public class ConfigImpl implements Config {
             builder.translation(translationKey("controller")).push("controller");
             energyCapacity = builder
                 .translation(translationKey("controller." + ENERGY_CAPACITY))
-                .defineInRange(ENERGY_CAPACITY, DefaultEnergyUsage.CONTROLLER_CAPACITY, 0, Long.MAX_VALUE);
+                .defineInRange(ENERGY_CAPACITY, DefaultEnergyUsage.CONTROLLER_CAPACITY, 1, Long.MAX_VALUE);
             builder.pop();
         }
 
@@ -912,7 +912,7 @@ public class ConfigImpl implements Config {
             builder.translation(translationKey("wirelessGrid")).push("wirelessGrid");
             energyCapacity = builder
                 .translation(translationKey("wirelessGrid." + ENERGY_CAPACITY))
-                .defineInRange(ENERGY_CAPACITY, DefaultEnergyUsage.WIRELESS_GRID_CAPACITY, 0, Long.MAX_VALUE);
+                .defineInRange(ENERGY_CAPACITY, DefaultEnergyUsage.WIRELESS_GRID_CAPACITY, 1, Long.MAX_VALUE);
             openEnergyUsage = builder
                 .translation(translationKey("wirelessGrid.openEnergyUsage"))
                 .defineInRange(OPEN_ENERGY_USAGE, DefaultEnergyUsage.WIRELESS_GRID_OPEN, 0, Long.MAX_VALUE);
@@ -982,7 +982,7 @@ public class ConfigImpl implements Config {
             builder.translation(translationKey("portableGrid")).push("portableGrid");
             energyCapacity = builder
                 .translation(translationKey("portableGrid." + ENERGY_CAPACITY))
-                .defineInRange(ENERGY_CAPACITY, DefaultEnergyUsage.PORTABLE_GRID_CAPACITY, 0, Long.MAX_VALUE);
+                .defineInRange(ENERGY_CAPACITY, DefaultEnergyUsage.PORTABLE_GRID_CAPACITY, 1, Long.MAX_VALUE);
             openEnergyUsage = builder
                 .translation(translationKey("portableGrid.openEnergyUsage"))
                 .defineInRange(OPEN_ENERGY_USAGE, DefaultEnergyUsage.PORTABLE_GRID_OPEN, 0, Long.MAX_VALUE);
@@ -1129,7 +1129,7 @@ public class ConfigImpl implements Config {
             builder.translation(translationKey("wirelessAutocraftingMonitor")).push("wirelessAutocraftingMonitor");
             energyCapacity = builder
                 .translation(translationKey("wirelessAutocraftingMonitor." + ENERGY_CAPACITY))
-                .defineInRange(ENERGY_CAPACITY, DefaultEnergyUsage.WIRELESS_AUTOCRAFTING_MONITOR_CAPACITY, 0,
+                .defineInRange(ENERGY_CAPACITY, DefaultEnergyUsage.WIRELESS_AUTOCRAFTING_MONITOR_CAPACITY, 1,
                     Long.MAX_VALUE);
             openEnergyUsage = builder
                 .translation(translationKey("wirelessAutocraftingMonitor.openEnergyUsage"))